+++ hypervisor/Makefile
@@ -221,10 +221,8 @@ C_SRCS += dm/vpci/vpci.c
 ifeq ($(CONFIG_PARTITION_MODE),y)
 C_SRCS += dm/vmptable.c
 endif
-C_SRCS += dm/vpci/partition_mode.c
 C_SRCS += dm/vpci/vhostbridge.c
 C_SRCS += dm/vpci/pci_pt.c
-C_SRCS += dm/vpci/sharing_mode.c
 C_SRCS += dm/vpci/vmsi.c
 C_SRCS += dm/vpci/vmsix.c
